Add Zettacache and slog to automation
This change allows for specifying a device to use as a Zettacache. The
device is partitioned into 8G for the slog and the remainder is used
as the Zettacache. If a device was specified, then create_pool will
automatically add the first partition as a log device.

Also, zfs_object_agent is now preferentially started via the service.