DOSE-725 zio read never returns when agent is restarted (#508)
If the agent restarts when there are outstanding "read block" requests,
but we are not in the middle of a txg, then it hangs.  The problem is
that the agent incorrectly thinks that we are in the middle of a txg,
and waits for a "resume done" message before allowing "read block"
requests to be processed.

The solution is for the kernel to tell the agent whether it's in the
middle of a txg as part of the "open pool" message.  Note that the old
"resume" argument (indicating if we are opening the pool for the first
time vs after agent restart) is not needed; these are handled
identically if we are not in the middle of a txg.

Since the resume_txg state is now communcated as part of the "open pool"
request, the "resume txg" request is no longer needed.