DOSE-745 DataObject::get_block() fails unwrap due to new object cache invalidation issue (#521)
If there's a concurrent get_object() while a put_object() completes, the
get_object may get the old value, which is fine.  But we can't allow new
readers to see the old value.  The problem is that the get_object() may
add the stale value to the cache, or pass it directly to other
concurrent get_object() calls that were initiated after the put_object()
completed.

The solution is for the put_object() to tell the get_object() to cancel
its caching/handoff of the potentially stale value.